requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
lxml==4.9.3
jinja2==3.1.2
pyyaml==6.0
